Move print_version_information() out of controller
