+++ newsboat.cpp
@@ -1,8 +1,11 @@
 #include <cstring>
 #include <errno.h>
 #include <iostream>
+#include <ncurses.h>
+#include <sys/utsname.h>
 
 #include "cache.h"
+#include "config.h"
 #include "cliargsparser.h"
 #include "config.h"
 #include "controller.h"
@@ -11,6 +14,7 @@
 #include "rss.h"
 #include "rsspp.h"
 #include "view.h"
+#include "xlicense.h"
 
 using namespace newsboat;
 
@@ -87,6 +91,58 @@ void print_usage(const char* argv0)
 	}
 }
 
+void print_version_information(const char* argv0, unsigned int level)
+{
+	if (level <= 1) {
+		std::cout << PROGRAM_NAME << " " << PROGRAM_VERSION << " - "
+			  << PROGRAM_URL << std::endl;
+		std::cout << "Copyright (C) 2006-2015 Andreas Krennmair"
+			  << std::endl;
+		std::cout << "Copyright (C) 2015-2018 Alexander Batischev"
+			  << std::endl;
+		std::cout << "Copyright (C) 2006-2017 Newsbeuter contributors"
+			  << std::endl;
+		std::cout << "Copyright (C) 2017-2018 Newsboat contributors"
+			  << std::endl;
+		std::cout << std::endl;
+
+		std::cout << strprintf::fmt(
+				     _("Newsboat is free software licensed "
+				       "under the MIT License. (Type `%s -vv' "
+				       "to see the full text.)"),
+				     argv0)
+			  << std::endl;
+		std::cout << _("It bundles JSON for Modern C++ library, "
+			       "licensed under the MIT License: "
+			       "https://github.com/nlohmann/json")
+			  << std::endl;
+		std::cout << std::endl;
+
+		struct utsname xuts;
+		uname(&xuts);
+		std::cout << PROGRAM_NAME << " " << PROGRAM_VERSION
+			  << std::endl;
+		std::cout << "System: " << xuts.sysname << " " << xuts.release
+			  << " (" << xuts.machine << ")" << std::endl;
+#if defined(__GNUC__) && defined(__VERSION__)
+		std::cout << "Compiler: g++ " << __VERSION__ << std::endl;
+#endif
+		std::cout << "ncurses: " << curses_version()
+			  << " (compiled with " << NCURSES_VERSION << ")"
+			  << std::endl;
+		std::cout << "libcurl: " << curl_version() << " (compiled with "
+			  << LIBCURL_VERSION << ")" << std::endl;
+		std::cout << "SQLite: " << sqlite3_libversion()
+			  << " (compiled with " << SQLITE_VERSION << ")"
+			  << std::endl;
+		std::cout << "libxml2: compiled with " << LIBXML_DOTTED_VERSION
+			  << std::endl
+			  << std::endl;
+	} else {
+		std::cout << LICENSE_str << std::endl;
+	}
+}
+
 int main(int argc, char* argv[])
 {
 	utils::initialize_ssl_implementation();
@@ -106,6 +162,10 @@ int main(int argc, char* argv[])
 
 	if (args.should_print_usage) {
 		print_usage(args.program_name.c_str());
+		return EXIT_SUCCESS;
+	} else if (args.show_version) {
+		print_version_information(
+			args.program_name.c_str(), args.show_version);
 		return EXIT_SUCCESS;
 	}
